fix: build scripts, Cargo profile, dev server config, and Vite optimization
- build-android: add incremental frontend detection (skip if unchanged)
- run-android: port 1420, add android init + patch steps
- run-windows: more graceful port cleanup
- Cargo.toml: tokio features reduced to rt+macros
- .cargo/config.toml: dev/release profile tuning
- tauri.android.conf.json: remove hardcoded devUrl
- vite.config.js: default port 1420, add manualChunks + dep pre-opt
- package.json: bump to 0.3.16, test:all -> run-all-tests.ps1
- news.rs: remove unnecessary as usize cast
- add run-vite-stub.bat, scripts/run-all-tests.ps1

+++ src-tauri/src/modules/news.rs
@@ -378,7 +378,7 @@ pub async fn fetch_news(db: &DatabasePool, region: &str, target_lang: &str) -> V
             if response.status().is_success() {
                 let body = response.text().await.unwrap_or_default();
                 if let Ok(feed) = feed_rs::parser::parse(body.as_bytes()) {
-                    for entry in feed.entries.iter().take(limit - raw_entries.len() as usize) {
+                    for entry in feed.entries.iter().take(limit - raw_entries.len()) {
                         let title = entry
                             .title
                             .as_ref()
